Retry on error
When making requests to the IATI Registry API, the registry refresher
will pause and retry up to five times if it receives a response status
code of 500 (Internal Server Error).

Currently, it is receiving response status codes of 503 (Service
Unavailable). It should pause and retry in these cases, too, but it
doesn’t. Ideally, it should retry if it receives a response status code
that is anything other than 2xx (success).

This change checks for a response status code of 200 (OK). If the
response differs from that, it will pause and then retry.